server/requirements.txt
mysql-connector-python
python-dotenv
flask
flask-cors
sqlalchemy
mysqlclient
pymysql
